qemu unit tests: use raw 128-bit asm syntax
The tests where developed using cpp macros to instanciate .insn
statements and shifts to retrieve the 64 upper bits of the registers
to check the 128-bit values.
Now that gas supports the 128-bit insns and constants and gdb outputs
128-bit registers values, we can use that for our tests.

Note that the status here is not at all operational, only (partial)
edition has been performed, and no assembly has ever been tried.
Do not expect to bissect with that