update: test can be compile with libfemto or newlib
add a new macro LIBFEMTO (defined in stdio) in orders to when this macro
is defined use print_uart to directly print elements on UART, else use
the printf of newlib.

remove stdnew.h and put definition of print_uart in stdio.

macro defined in stdnew is now in utils/headers/global_var.h

update: execute-test to execute in newlib or libfemto mode

add: compile_newlib.sh a script that run makefiles